Infer pointer-backed structures from fixed offsets

+++ decomp/ST.exe/functions/00417A20/decomp.c
@@ -15,7 +15,7 @@ undefined4 __thiscall FUN_00417a20(void *this,short param_1,short param_2,short 
     if (param_4 == 1) {
       thunk_FUN_00495ff0(*(short *)((int)this + 0x5b),*(short *)((int)this + 0x5d),
                          *(short *)((int)this + 0x5f),
-                         CONCAT31(uVar3,*(undefined1 *)((int)this + 0x8e)),(int)this);
+                         CONCAT31(uVar3,*(undefined1 *)((int)this + 0x8e)),this);
     }
     *(short *)((int)this + 0x47) = param_1;
     *(short *)((int)this + 0x4b) = param_3;
@@ -27,7 +27,7 @@ undefined4 __thiscall FUN_00417a20(void *this,short param_1,short param_2,short 
       uVar1 = DumpClassC::WritePtr
                         (param_1,param_2,param_3,
                          CONCAT31((int3)((uint)*(undefined4 *)((int)this + 0x18) >> 8),
-                                  *(undefined1 *)((int)this + 0x8e)),(int)this);
+                                  *(undefined1 *)((int)this + 0x8e)),this);
       return uVar1;
     }
   }
@@ -35,19 +35,18 @@ undefined4 __thiscall FUN_00417a20(void *this,short param_1,short param_2,short 
     if (param_4 == 1) {
       uVar1 = thunk_FUN_00495ff0(*(short *)((int)this + 0x5b),*(short *)((int)this + 0x5d),
                                  *(short *)((int)this + 0x5f),
-                                 CONCAT31(uVar3,*(undefined1 *)((int)this + 0x8e)),(int)this);
+                                 CONCAT31(uVar3,*(undefined1 *)((int)this + 0x8e)),this);
       thunk_FUN_00495ff0(*(short *)((int)this + 0x5b) + 1,*(short *)((int)this + 0x5d),
                          *(short *)((int)this + 0x5f),
-                         CONCAT31((int3)((uint)uVar1 >> 8),*(undefined1 *)((int)this + 0x8e)),
-                         (int)this);
+                         CONCAT31((int3)((uint)uVar1 >> 8),*(undefined1 *)((int)this + 0x8e)),this);
       thunk_FUN_00495ff0(*(short *)((int)this + 0x5b),*(short *)((int)this + 0x5d) + 1,
                          *(short *)((int)this + 0x5f),
                          CONCAT31((int3)((uint)extraout_EDX >> 8),*(undefined1 *)((int)this + 0x8e))
-                         ,(int)this);
+                         ,this);
       thunk_FUN_00495ff0(*(short *)((int)this + 0x5b) + 1,*(short *)((int)this + 0x5d) + 1,
                          *(short *)((int)this + 0x5f),
                          CONCAT31((int3)((uint)extraout_ECX >> 8),*(undefined1 *)((int)this + 0x8e))
-                         ,(int)this);
+                         ,this);
     }
     *(short *)((int)this + 0x47) = param_1;
     *(short *)((int)this + 0x4b) = param_3;
@@ -59,7 +58,7 @@ undefined4 __thiscall FUN_00417a20(void *this,short param_1,short param_2,short 
       iVar2 = DumpClassC::WritePtr
                         (param_1,param_2,param_3,
                          CONCAT31((int3)((uint)*(undefined4 *)((int)this + 0x18) >> 8),
-                                  *(undefined1 *)((int)this + 0x8e)),(int)this);
+                                  *(undefined1 *)((int)this + 0x8e)),this);
       if (iVar2 != 0) {
         return 0xffffffff;
       }
@@ -67,13 +66,13 @@ undefined4 __thiscall FUN_00417a20(void *this,short param_1,short param_2,short 
                         (*(short *)((int)this + 0x47) + 1,*(short *)((int)this + 0x49),
                          *(short *)((int)this + 0x4b),
                          CONCAT31((int3)((uint)extraout_ECX_00 >> 8),
-                                  *(undefined1 *)((int)this + 0x8e)),(int)this);
+                                  *(undefined1 *)((int)this + 0x8e)),this);
       if (iVar2 != 0) {
         return 0xffffffff;
       }
       iVar2 = DumpClassC::WritePtr
                         (*(short *)((int)this + 0x47),*(short *)((int)this + 0x49) + 1,
-                         *(short *)((int)this + 0x4b),(uint)*(byte *)((int)this + 0x8e),(int)this);
+                         *(short *)((int)this + 0x4b),(uint)*(byte *)((int)this + 0x8e),this);
       if (iVar2 != 0) {
         return 0xffffffff;
       }
@@ -81,7 +80,7 @@ undefined4 __thiscall FUN_00417a20(void *this,short param_1,short param_2,short 
                         (*(short *)((int)this + 0x47) + 1,*(short *)((int)this + 0x49) + 1,
                          *(short *)((int)this + 0x4b),
                          CONCAT31((int3)((uint)extraout_EDX_00 >> 8),
-                                  *(undefined1 *)((int)this + 0x8e)),(int)this);
+                                  *(undefined1 *)((int)this + 0x8e)),this);
       return uVar1;
     }
   }
