+++ decomp/ST.exe/functions/004C91A0/decomp.c
@@ -10,7 +10,7 @@ undefined4 __fastcall FUN_004c91a0(void *param_1)
   if (*(int *)((int)param_1 + 0x361) != 0) {
     *(undefined4 *)((int)param_1 + 0x36d) = 1;
     *(int *)((int)param_1 + 0x365) = *(int *)((int)param_1 + 0x361);
-    iVar1 = thunk_FUN_004c90c0((int)param_1);
+    iVar1 = thunk_FUN_004c90c0(param_1);
     if (iVar1 != 0) {
       iVar1 = thunk_FUN_004406c0(*(char *)((int)param_1 + 0x24));
       if (((char)iVar1 != '\x03') && (*(int *)((int)param_1 + 0x398) != 0)) {
