+++ decomp/ST.exe/functions/004D0970/decomp.c
@@ -19,7 +19,7 @@ FUN_004d0970(void *this,undefined4 param_1,undefined4 param_2,undefined4 param_3
   *(undefined4 *)(*(int *)((int)this + 0x24d) + *(int *)((int)this + 0x245) * 0xc) = param_1;
   *(undefined4 *)(*(int *)((int)this + 0x24d) + 4 + *(int *)((int)this + 0x245) * 0xc) = param_2;
   *(undefined4 *)(*(int *)((int)this + 0x24d) + 8 + *(int *)((int)this + 0x245) * 0xc) = param_3;
-  iVar1 = DumpClassC::WritePtr((short)param_1,(short)param_2,(short)param_3,0,(int)this);
+  iVar1 = DumpClassC::WritePtr((short)param_1,(short)param_2,(short)param_3,0,this);
   if (iVar1 != 0) {
     RaiseInternalException
               (iVar1,g_overwriteContext_007ED77C,s_E____titans_Artem_TLO_dock_cpp_007bf39c,0x2f9);
