+++ decomp/ST.exe/functions/006DD800/decomp.c
@@ -12,7 +12,7 @@ void __thiscall FUN_006dd800(void *this,float param_1)
          (double)((float)_DAT_0079df78 / (param_1 - (float)*(double *)((int)this + 0x114)));
     lVar1 = Library::MSVCRT::__ftol();
     *(int *)((int)this + 0x2e4) = (int)lVar1;
-    FUN_006dd870((int)this);
+    FUN_006dd870(this);
     return;
   }
   return;
