+++ decomp/ST.exe/functions/0074BC50/decomp.c
@@ -3,7 +3,7 @@ void FUN_0074bc50(LPVOID param_1)
 
 {
   if (param_1 != (LPVOID)0x0) {
-    FUN_0074bce4((int)param_1);
+    FUN_0074bce4(param_1);
     CoTaskMemFree(param_1);
   }
   return;
