+++ decomp/ST.exe/functions/0075034D/decomp.c
@@ -6,7 +6,7 @@ undefined4 __thiscall FUN_0075034d(void *this,undefined4 param_1,undefined4 para
   *(undefined4 *)((int)this + 0x68) = param_1;
   *(undefined4 *)((int)this + 0x6c) = param_2;
   *(undefined4 *)((int)this + 100) = 1;
-  FUN_00750287((int)this);
+  FUN_00750287(this);
   LeaveCriticalSection((LPCRITICAL_SECTION)((int)this + 8));
   return 0;
 }
